update terrible windows settings paths, remove unused event handler code, add warning flags, change audio buffer code, update readme

+++ ext/include/clay_renderer_SDL3.hpp
@@ -44,7 +44,7 @@ struct CameraData {
     // internal
     void* __pixels;
     size_t __pixelsSize;
-    int __pitch;
+    size_t __pitch;
 
     SDL_FRect __prevRect;
     CameraDisplayMode __prevDisplayMode = static_cast<CameraDisplayMode>(-1);
